+++ lib/rubocop/rspec/inject.rb
@@ -6,9 +6,11 @@ module RuboCop
     # bit of our configuration.
     module Inject
       def self.defaults!
-        path = CONFIG_DEFAULT.to_s
+        project_root = Pathname.new(__dir__).parent.parent.parent.expand_path
+        config_default = project_root.join('config', 'default.yml')
+        path = config_default.to_s
         hash = ConfigLoader.send(:load_yaml_configuration, path)
-        config = Config.new(hash, path)
+        config = RuboCop::Config.new(hash, path)
         puts "configuration from #{path}" if ConfigLoader.debug?
         config = ConfigLoader.merge_with_default(config, path)
         ConfigLoader.instance_variable_set(:@default_configuration, config)